% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageProfile.R
\name{averageProfile}
\alias{averageProfile}
\title{averageProfile to get profile from RNMT-enriched targets and not enriched targets.
Pre-request: download ngs.plot from https://drive.google.com/drive/folders/0B1PVLadG_dCKN1liNFY0MVM1Ulk}
\usage{
averageProfile(sampleFiles, region = "genebody",
  mergeFile = "merged.bam", mergeNGSFile = "mergedNGS.bam",
  enrichedFile = "enriched.txt", unenrichedFile = "unenriched.txt")
}
\arguments{
\item{sampleFiles}{The samples of alignment files as a list in bam form}

\item{region}{genome region to plot, should be tss, tes, genebody, exon, cgi, enhancer, dhs or bed}

\item{mergeFile}{File name of merged file}

\item{mergeNGSFile}{File name of merged file with all start with chr, only bam file with same pattern could use ngs.plot}

\item{enrichedFile}{File name of genes with enriched RNMT}

\item{unenrichedFile}{File name of genes with unenriched RNMT}
}
\description{
\code{averageProfile} gets profile from RNMT-enriched target (50 counts per million), the aligned region could be tss, tes, genebody, exon, cgi, enhancer, dhs or bed.
}
\examples{
\dontrun{
sampleFiles <- c("../BCBProjectData/SRR4393138.Aligned.sortedByCoord.out.bam",
"../BCBProjectData/SRR4393139.Aligned.sortedByCoord.out.bam",
"../BCBProjectData/SRR4393140.Aligned.sortedByCoord.out.bam")
averageProfile(sampleFiles)
}
}
