% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTSS.R
\name{expandTSS}
\alias{expandTSS}
\title{Expand TSS (transcript start site)}
\usage{
expandTSS(outputFile)
}
\arguments{
\item{outputFile}{the name of output file, default is TSSextend.bed}
}
\description{
\code{expandTSS} Expand the TSS to 500
}
\examples{
\dontrun{
 expandTSS()
}
}
