% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeCoverage.R
\name{genomeCoverage}
\alias{genomeCoverage}
\title{Gene coverage of samples over annotated regions of genome.}
\usage{
genomeCoverage(sampleFiles, annotaedGenomes, UTRDiffL = NULL,
  mergeFile = "merged.bam")
}
\arguments{
\item{sampleFiles}{The samples of alignment files as a list in bam form}

\item{annotaedGenomes}{The list of annotated genomes for "5' UTR", "3' UTR", "exons", "introns", "promoters", "lncRNA", "tRNA", "snc / miRNA". The order of the annotated genome is important.}

\item{UTRDiffL}{UTR feature file with different length in bed format, default if NULL}

\item{mergeFile}{default "merged.bam". Merged file of all samples to do gene coverage}
}
\description{
\code{genomeCoverage} Plot the genome coverage plot for 5'UTR, 3'UTR, exons, introns, promoters. lncRNA, sno/miRNA.
}
\examples{
\dontrun{
sampleFiles <- c("../BCBProjectData/SRR4393138.Aligned.sortedByCoord.out.bam",
"../BCBProjectData/SRR4393139.Aligned.sortedByCoord.out.bam",
"../BCBProjectData/SRR4393140.Aligned.sortedByCoord.out.bam")
annotaedGenomes <- c("../BCBProjectData/5UTR2000.bed", "../BCBProjectData/3UTR2000.bed",
"../BCBProjectData/GENCODE.Exon.bed", "../BCBProjectData/GENCODE.Intron.bed",
"../BCBProjectData/GENCODE.Promoter.bed", "../BCBProjectData/lincRNA.bed",
"../BCBProjectData/tRNA.bed", "../BCBProjectData/miRNA.bed")
UTRDiffL <- c("../BCBProjectData/3UTR2000.bed","../BCBProjectData/3UTR1000.bed",
"../BCBProjectData/3UTR500.bed","../BCBProjectData/3UTR200.bed",
"../BCBProjectData/5UTR2000.bed","../BCBProjectData/5UTR1500.bed",
"../BCBProjectData/5UTR1000.bed","../BCBProjectData/5UTR500.bed")
genomeCoverage(sampleFiles, annotaedGenomes, UTRDiffL)
}
}
