% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTSS.R
\name{getTSS}
\alias{getTSS}
\title{Get TSS (transcript start site)}
\usage{
getTSS(fileName, outputFile = "TSS.bed")
}
\arguments{
\item{fileName}{the name of the file a gtf file}

\item{outputFile}{the name of output file, default is TSS.bed}
}
\description{
\code{getTSS} Find TSS sites and store in bed file (tab seperated)
}
\examples{
\dontrun{
 getTSS("../BCBProjectData/gencode.v19.annotation.gtf")
}
}
\seealso{
\code{\link[rtracklayer]{import}} is used to import the gtf file.
}
