% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpTPM.R
\name{SpTPM}
\alias{SpTPM}
\title{Scatterplot of TPM of inputs vs. average TPM of sample.}
\usage{
SpTPM(inputFile, sampleFiles, version = "hg19")
}
\arguments{
\item{inputFile}{The input of alignment files as a list in bam form}

\item{sampleFiles}{The samples of alignment files as a list in bam form}

\item{version}{default is "hg19". Could be "mm10", "mm9", "hg38" and "hg19"}
}
\description{
\code{SpTPM} Scatterplot of TPM of inputs vs. average TPM of sample. Count reads per trancript and get the TPM; plot the scatterplot for average TPM samples and TPM inputs.
}
\examples{
\dontrun{
inputFile <- "../BCBProjectData/SRR4393137.Aligned.sortedByCoord.out.bam"
sampleFiles <- c("../BCBProjectData/SRR4393138.Aligned.sortedByCoord.out.bam",
"../BCBProjectData/SRR4393139.Aligned.sortedByCoord.out.bam",
"../BCBProjectData/SRR4393140.Aligned.sortedByCoord.out.bam")
SpTPM(inputFile, sampleFiles)
}
}
